Add and update PCSX2 patches for multiple games
Added new 60 FPS and 50/60 FPS patches for Scooby-Doo! Night of 100 Frights (PAL-E and NTSC-U), Harry Potter and the Chamber of Secrets (PAL-E), Shadow Man - 2econd Coming (NTSC-U), 007 - Nightfire (NTSC-U), and Crash of the Titans (NTSC-U). Updated the 50/60 FPS patch for Scarface - The World is Yours (PAL-M) with a new code.